+++ src/components/chat/Content/index.tsx
@@ -1,0 +1,7 @@
+export default function Content() {
+  return (
+    <div className={'flex-1'}>
+      content
+    </div>
+  )
+}